add github actions
having the final pdfs in the repo is like having the final binaries in the repo. it's not a good idea. this PR adds a github action that builds the pdfs and uploads them as artifacts. the artifacts are available for download from the github ui.
furthermore, the action also creates a release with the pdfs as assets. this way, the pdfs are available for download from the releases page as well.