added example slides to README.md
That way, users get an impression of the final result as soon as they open the repo, without having to download any PDF files from the releases page.

The fact that complete PDFs can be found on the releases page is also mentioned.

Furthermore, instructions on how to build new PDFs in the GitHub/GitLab CI were moved to the end.